backend/requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==23.5
requests==2.31.0
pandas==2.2.2
openpyxl==3.1.2